scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.1",
    "opencv-python>=4.11.0.86",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.5",
    "easyocr>=1.7.2",
    "beautifulsoup4>=4.12.3",
]
